pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gossiptoon"
version = "0.1.0"
description = "AI-powered video generation system that converts Reddit gossip into YouTube Shorts"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "GossipToon Team", email = "team@gossiptoon.dev"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Video",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    # Core framework
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",

    # LangChain ecosystem
    "langchain>=0.3.15",
    "langchain-core>=0.3.28",
    "langchain-openai>=0.2.14",
    "langgraph>=0.2.59",

    # LLM providers
    "openai>=1.59.5",
    "google-generativeai>=0.8.3",

    # Audio
    "elevenlabs>=1.15.0",
    "openai-whisper>=20240930",
    "pydub>=0.25.1",

    # Reddit/Web search
    "praw>=7.7.1",
    "tavily-python>=0.5.0",

    # Utilities
    "requests>=2.32.3",
    "httpx>=0.28.1",
    "tenacity>=9.0.0",
    "rich>=13.9.4",
    "click>=8.1.7",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",

    # Linting & Formatting
    "ruff>=0.8.4",
    "black>=24.10.0",
    "mypy>=1.13.0",

    # Type stubs
    "types-PyYAML>=6.0.12.20240917",
    "types-requests>=2.32.0.20241016",
]

[project.scripts]
gossiptoon = "gossiptoon.main:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=gossiptoon",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false

[tool.black]
line-length = 100
target-version = ["py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py312"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = false
disallow_subclassing_any = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
check_untyped_defs = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "elevenlabs.*",
    "whisper.*",
    "praw.*",
    "tavily.*",
]
ignore_missing_imports = true
